Combine `filter` and `filter_indexed`
