+++ docs/docs.md
@@ -16,7 +16,6 @@
  * [each](functions/each.md): Apply a function over all the items of a collection
  * [filter](functions/filter.md): Discriminate the items of a collection for which function is false
  * [filter_fresh](functions/filter_fresh.md): Similar to filter, but returns a collection that starts at 0
- * [filter_indexed](functions/filter_indexed.md): Discriminate the items of a collection for which function is false using also the key
  * [filter_null](functions/filter_null.md): Discriminate the items of a collection for which value is null
  * [first](functions/first.md): Returns the first element of a collection
  * [flat_map](functions/flat_map.md): Returns an array containing the results of applying a given function to the items of a collection and flattening the results
