+++ src/_bootstrap.php
@@ -16,7 +16,6 @@ require __DIR__ . '/do_if.php';
 require __DIR__ . '/each.php';
 require __DIR__ . '/filter.php';
 require __DIR__ . '/filter_fresh.php';
-require __DIR__ . '/filter_indexed.php';
 require __DIR__ . '/filter_null.php';
 require __DIR__ . '/first.php';
 require __DIR__ . '/flat_map.php';
