--- src/filter_indexed.php
@@ -1,29 +1,0 @@
-<?php
-
-declare(strict_types = 1);
-
-namespace Lambdish\Phunctional;
-
-/**
- * Returns an array with the items in $coll for which $fn returns true.
- *
- * Similar to `array_filter` but with a consistent parameters order, requiring always a function and allowing access
- * to the keys of the collection.
- *
- * @param callable $fn   function to filter by
- * @param iterable $coll collection of values to be filtered
- */
-function filter_indexed(callable $fn, iterable $coll): array
-{
-    $result = [];
-
-    foreach ($coll as $key => $value) {
-        if ($fn($value, $key)) {
-            $result[$key] = $value;
-        }
-    }
-
-    return $result;
-}
-
-const filter_indexed = '\Lambdish\Phunctional\filter_indexed';
